rename packages to io.cdap and update to 6.0.0-snapshot
