+++ src/test/java/io/cdap/plugin/filecontent/FileContentsActionTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2017 Cask Data, Inc.
+ * Copyright © 2017-2019 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,9 +14,9 @@
  * the License.
  */
 
-package co.cask.hydrator.plugin;
+package io.cdap.plugin.filecontent;
 
-import co.cask.cdap.etl.mock.common.MockPipelineConfigurer;
+import io.cdap.cdap.etl.mock.common.MockPipelineConfigurer;
 import org.junit.Test;
 
 import java.io.File;
